@@ -1,6 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
     <PropertyGroup>
-        <TargetFramework>net6.0</TargetFramework>
+        <TargetFramework>net7.0</TargetFramework>
         <AppendTargetFrameworkToOutputPath>false</AppendTargetFrameworkToOutputPath>
         <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
         <LangVersion>10.0</LangVersion>
@@ -28,17 +28,17 @@
         <WarningLevel>4</WarningLevel>
     </PropertyGroup>
     <ItemGroup>
-        <PackageReference Include="FluentValidation" Version="8.6.3"/>
-        <PackageReference Include="Microsoft.AspNetCore.Owin" Version="6.0.5"/>
-        <PackageReference Include="Microsoft.Extensions.Hosting" Version="6.0.1"/>
-        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="6.0.0"/>
-        <PackageReference Include="Microsoft.Owin.SelfHost" Version="4.2.2"/>
-        <PackageReference Include="Nancy" Version="2.0.0"/>
-        <PackageReference Include="Nancy.Authentication.Basic" Version="2.0.0"/>
-        <PackageReference Include="Nancy.Owin" Version="2.0.0"/>
-        <PackageReference Include="Nancy.Serialization.JsonNet" Version="2.0.0-clinteastwood"/>
-        <PackageReference Include="Nancy.Validation.FluentValidation" Version="2.0.0"/>
-        <PackageReference Include="Newtonsoft.Json" Version="13.0.1"/>
-        <PackageReference Include="System.Configuration.ConfigurationManager" Version="6.0.0"/>
+        <PackageReference Include="FluentValidation" Version="8.6.3" />
+        <PackageReference Include="Microsoft.AspNetCore.Owin" Version="7.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Hosting" Version="7.0.0" />
+        <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="7.0.0" />
+        <PackageReference Include="Microsoft.Owin.SelfHost" Version="4.2.2" />
+        <PackageReference Include="Nancy" Version="2.0.0" />
+        <PackageReference Include="Nancy.Authentication.Basic" Version="2.0.0" />
+        <PackageReference Include="Nancy.Owin" Version="2.0.0" />
+        <PackageReference Include="Nancy.Serialization.JsonNet" Version="2.0.0-clinteastwood" />
+        <PackageReference Include="Nancy.Validation.FluentValidation" Version="2.0.0" />
+        <PackageReference Include="Newtonsoft.Json" Version="13.0.2" />
+        <PackageReference Include="System.Configuration.ConfigurationManager" Version="7.0.0" />
     </ItemGroup>
 </Project>
