@@ -1,12 +1,12 @@
-﻿using System;
-using System.Configuration;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Nancy.Owin;
+using System;
+using System.Configuration;
+using System.Threading.Tasks;
 
 namespace Building.A.RESTful.API.With.NancyFx;
 
@@ -14,8 +14,8 @@ internal class Program
 {
     private static void PopulateRepositories()
     {
-        PersonRepository.Instance.Add(new Person {Id = 1, FirstName = "Danut", LastName = "Radoaica"});
-        UserRepository.Instance.Add(new User {UserName = "dradoaica", Password = "1234", Identifier = Guid.NewGuid()});
+        PersonRepository.Instance.Add(new Person { Id = 1, FirstName = "Danut", LastName = "Radoaica" });
+        UserRepository.Instance.Add(new User { UserName = "dradoaica", Password = "1234", Identifier = Guid.NewGuid() });
     }
 
     private static void Main()
@@ -26,19 +26,19 @@ internal class Program
     private static async Task MainAsync()
     {
         _ = ConfigurationManager.AppSettings.AllKeys;
-        var url = "http://+:8984";
+        string url = "http://+:8984";
         IHostBuilder hostBuilder = new HostBuilder()
             .UseEnvironment(Environments.Development)
             .ConfigureServices(services =>
             {
-                services.Configure<KestrelServerOptions>(options => { options.AllowSynchronousIO = true; });
-                services.Configure<IISServerOptions>(options => { options.AllowSynchronousIO = true; });
+                _ = services.Configure<KestrelServerOptions>(options => { options.AllowSynchronousIO = true; });
+                _ = services.Configure<IISServerOptions>(options => { options.AllowSynchronousIO = true; });
                 PopulateRepositories();
             })
             .ConfigureWebHostDefaults(webBuilder =>
             {
-                webBuilder.UseUrls(url);
-                webBuilder.UseStartup<Startup>();
+                _ = webBuilder.UseUrls(url);
+                _ = webBuilder.UseStartup<Startup>();
             });
         IHost host = hostBuilder.Build();
         using (host)
@@ -51,13 +51,13 @@ internal class Program
     {
         public void Configure(IApplicationBuilder app)
         {
-            app.UseOwin(pipeline =>
+            _ = app.UseOwin(pipeline =>
             {
-                var options = new NancyOptions
+                NancyOptions options = new()
                 {
                     Bootstrapper = new Bootstrapper()
                 };
-                pipeline.UseNancy(options);
+                _ = pipeline.UseNancy(options);
             });
         }
     }
