@@ -1,6 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using Nancy.Authentication.Basic;
+using System.Collections.Generic;
 using System.Security.Claims;
-using Nancy.Authentication.Basic;
 
 namespace Building.A.RESTful.API.With.NancyFx;
 
@@ -11,9 +11,13 @@ public class UserValidator : IUserValidator
         // null == anonymous
         ClaimsPrincipal ret = null;
         foreach (User user in UserRepository.Instance.GetAll())
+        {
             if (user.UserName == username && user.Password == password)
+            {
                 ret = new UserIdentity(user.UserName, user.Identifier,
-                    new List<string> {UserIdentity.HANDLEPERSON_PERMISSION});
+                    new List<string> { UserIdentity.HANDLEPERSON_PERMISSION });
+            }
+        }
 
         return ret;
     }
